Simplified the controller functions
The three functions related to updating the arrays of buttons tates are now a single function, updateButtons.

The three functions isPressed(), isHeld(), and isReleased() have been merged into a single function, checkButton() which takes parameters for which button to check and what state or event (PRESS, HOLD, RELEASE) to check for.